requirements.txt
Django>=4.2.0,<5.3.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-filter>=23.2
Pillow>=9.5.0
SQLAlchemy>=1.4
Faker>=10.0
requests>=2.31.0
openai>=1.0.0
gunicorn>=21.2.0
django-cors-headers>=4.0.0
drf-spectacular>=0.26.0
python-dotenv>=1.0.0
celery>=5.3.0
redis>=4.5.0